Add test for building a pattern with guard.
For @OvermindDL1 , this feature is just experimental, haha.

Patterns without guards can be used as data constructors when all
its variables have been filled. Not so with guarded patterns, because
when you expand `t2(a: 1, b: 2)` having defined

```elixir
defpat t2({a, b}) when a > 1
```

It will expand into:

```elixir
{a = 1, b = 2} when a > b
```

And it however is not a valid expression outside of the context
of a function def, a case, with etc.

This patch implements a custom `build` option to get around this.
Take a look at the foo_test.exs.